session: Cache only when `should_cache` arg is true, otherwise clear it
Previously caching was rudimentarily implemented using a `lru_cache`
with a maxsize of 1024. That implementation, however, encouraged
re-requesting the same URLs even where it is avoidable, which can
cause bugs.

`should_cache` should be passed as `True` during forum software
detection and board hierarchy extraction. During retrieval of threads
and posts it should be `False` (default)